refactor(object-storage): move all endpoint to separated controller
